Move more specific code out of convert.go: restart policy
